move CuArrays to optional deps
